requirements.txt
requests
buildozer
kivy
kivymd
cython
setuptools
py-androidbuild
plyer
pyjnius

yt-dlp
# pip install -U yt-dlp
# python -m pywin32_postinstall -install
# python -m pip install --upgrade pywin32
